requirements.txt
sqlalchemy
python-dotenv
requests
pandas
pymysql
matplotlib
seaborn
mlflow
statsmodels
scikit-learn
streamlit
huggingface-hub